backend/requirements.txt
uvicorn==0.34.0
python-keycloak==5.1.1
PyJWT==2.10.1
fastapi[standard]==0.115.6
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.30
langchain-openai==0.3.0
langchain-postgres==0.0.12
langchain-text-splitters==0.3.5
opentelemetry-instrumentation-langchain==0.36.0
traceloop-sdk==0.36.0
langgraph==0.2.63